requirements.txt
flask-sqlalchemy 
flask_restful 
pymysql
itsdangerous==1.1.0